Remove tabindex from tab bar
Fixes atom/atom#6662
Refs #124

This was introduced in 310c058b6a61348f3962ec0a47a683e20784f4cd, with
a description that it prevented focus from being stolen when the close
icon is clicked. I don’t really understand the reasoning there, and
removing it fixes the above issue and doesn’t seem to introduce any
problems with focus on clicking the close icon.

@kevinsawicki perhaps you can comment on whether this makes sense?